Combine all code into 1 file
