+++ sorting/multi_threaded/merge_sort/multi_threaded_merge_sort.go
@@ -1,5 +1,10 @@
 package main
 
+import (
+	"fmt"
+	"math/rand"
+	"time"
+)
 
 const threshold = 10000
 
@@ -53,3 +58,89 @@ func merge(leftArray, rightArray []int) []int {
 	}
 	return result
 }
+
+func mergeSort(inputArray []int) []int {
+	if len(inputArray) <= 1 {
+		return inputArray
+	}
+
+	middleIndex := len(inputArray) / 2
+	leftArray := mergeSort(inputArray[:middleIndex])
+	rightArray := mergeSort(inputArray[middleIndex:])
+
+	return merge(leftArray, rightArray)
+}
+
+func timSort(array []int) {
+	arrayLen := len(array)
+	minRun := 32
+
+	// Sort the small runs using insertion sort
+	for i := 0; i < arrayLen; i += minRun {
+		end := min(i+minRun, arrayLen)
+		insertionSort(array[i:end])
+	}
+
+	// Merge the sorted runs using merge sort
+	currentRunSize := minRun
+	for currentRunSize < arrayLen {
+		for leftStart := 0; leftStart < arrayLen; leftStart += 2 * currentRunSize {
+			middleIndex := min(leftStart+currentRunSize, arrayLen)
+			rightEnd := min(leftStart+2*currentRunSize, arrayLen)
+
+			leftArray := array[leftStart:middleIndex]
+			rightArray := array[middleIndex:rightEnd]
+			mergedArray := merge(leftArray, rightArray)
+
+			copy(array[leftStart:rightEnd], mergedArray)
+		}
+		currentRunSize *= 2
+	}
+}
+
+func insertionSort(array []int) {
+	n := len(array)
+	for i := 1; i < n; i++ {
+		key := array[i]
+		j := i - 1
+		for j >= 0 && key < array[j] {
+			array[j+1] = array[j]
+			j--
+		}
+		array[j+1] = key
+	}
+}
+
+func main() {
+	arraySize := 100000
+
+	array1 := []int{}
+
+	for i := 0; i < arraySize; i++ {
+		array1 = append(array1, rand.Intn(arraySize))
+	}
+
+	// Timing normal merge sort
+	startMergeSort := time.Now()
+	mergeSort(array1)
+	elapsedMergeSort := time.Since(startMergeSort)
+	fmt.Println("MergeSort for", arraySize, "elements took", elapsedMergeSort)
+
+	// Timing parallel merge sort
+	array2 := make([]int, len(array1))
+	copy(array2, array1)
+	startParallelMergeSort := time.Now()
+	ch := make(chan []int)
+	go parallelMergeSort(array2, ch)
+	<-ch
+	elapsedParallelMergeSort := time.Since(startParallelMergeSort)
+	fmt.Println("ParallelMergeSort for", arraySize, "elements took", elapsedParallelMergeSort)
+
+	// Compare the two
+	speedupPercentage := (float64(elapsedMergeSort-elapsedParallelMergeSort) / float64(elapsedMergeSort)) * 100
+	fmt.Println("ParallelMergeSort was", speedupPercentage, "percent faster")
+}
+
+
+// Result:
+// ParallelMergeSort was 63.96576490304046 percent faster
